+++ client/src/app/shared/components/loader-v1/loader-v1.component.ts
@@ -1,0 +1,14 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-loader-v1',
+  templateUrl: './loader-v1.component.html'
+})
+export class LoaderV1Component implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}