+++ client/src/app/shared/components/preloader-v1/preloader-v1.component.ts
@@ -1,0 +1,14 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-preloader-v1',
+  templateUrl: './preloader-v1.component.html'
+})
+export class PreloaderV1Component implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}